[Opcode/Struct/Build]: SMSG_INIT_CURRENCY + SMSG_UPDATE_CURRENCY + SMSG_UPDATE_CURRENCY_WEEK_LIMIT + SMSG_ARENA_SEASON_WORLDSTATE + SMSG_HOTFIX_INFO + SMSG_SET_FORCED_REACTIONS + SMSG_SET_FACTION_STANDING

+++ src/server/scripts/Commands/cs_debug.cpp
@@ -537,7 +537,7 @@ public:
 
     static bool HandleDebugSendQuestPartyMsgCommand(ChatHandler* handler, char const* args)
     {
-        uint32 msg = atol((char*)args);
+        uint8 msg = atol((char*)args);
         handler->GetSession()->GetPlayer()->SendPushToPartyResponse(handler->GetSession()->GetPlayer(), msg);
         return true;
     }
